Remove output check for ExampleTxn_Mutate_list.
Why is this change necessary?

This test was failing:

    --- FAIL: ExampleTxn_Mutate_list (0.01s)
    got:
    {"me":[{"address":["Riley Street","Redfern"],"phone_number":[123,9876]}]}
    want:
    {"me":[{"address":["Riley Street","Redfern"],"phone_number":[9876,123]}]}

List types are unordered sets and the particular order of the values
is non-deterministic, so the result of the test is OK.

How does it address the issue?

The output check is skipped by removing the "Output: " prefix in the comment.