Add drop_data field to api to drop data except schema and types.
This change adds API support for a new operation called drop_data which
will drop all the data but not the schema nor the types. This is useful
when the schema will be resued after dropping the data and the schema is
large enough that recreating it might take more than a couple seconds.