Set explicit Uid field for JSON mutation in example test. (#85)
This fixes an example that used the auto-generated blank node
name "blank-0". Auto-generated blank node names have changed with
dgraph-io/dgraph#3795. The test now sets the blank node name and
uses that instead of relying on Dgraph's auto-generated blank
node name in the mutation response.

Note: This example isn't configured to run as a Go test because
"// Output:" is intentionally omitted for this example. As
indicated in the comment above the last line, the output is
considered unordered for list schema types. This could be made as
a runnable test by printing out the individual edge values and
using the "Unordered output:" check instead of the raw
JSON response.