Add schema information to mutation message.
JSON mutations have more context about the type of the predicates. For
example, if the values for the predicate <name> are sent in a list, the
predicate <name> should be of type [string]. This information is lost
when JSON is converted into NQuads. The added fields will be used to
convey the previously lost information for further processing.